requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
pydantic-settings>=2.9.1
python-dotenv>=1.0.0

# LLM and Embeddings
langchain>=0.0.350
langchain-openai>=0.0.2
langgraph>=0.0.15
transformers>=4.36.2
torch>=2.1.2
accelerate>=0.25.0
sentence-transformers>=2.2.2

# Document Processing
pymupdf>=1.23.8  # PDF processing
python-docx>=1.0.1
striprtf>=0.0.26
pandas>=2.1.3
openpyxl>=3.1.2
python-pptx>=0.6.22
beautifulsoup4>=4.12.2
lxml>=4.9.3
paddleocr>=2.7.0.3
pytesseract>=0.3.10
spacy>=3.7.2
PyYAML>=6.0.1

# Image Processing
Pillow>=10.1.0
opencv-python>=4.8.1.78

# Database and Caching
redis>=5.0.1
psycopg2-binary>=2.9.9
chromadb>=0.4.18

# Utilities
tqdm>=4.66.1
numpy>=1.26.2
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
